fix(ci): Use GitHub Action home variables for priming cache
The whole reason for typesetting `cache.tex` is that we get a primed
cache, thus subsequent builds using the template should be much
faster. However, according to the [official documentation] GitHub
Actions set a custom home directory, effectively discarding our primed
cache and forcing tectonic to start from scratch.

This commit sets the HOME variable to the same directory as GitHub
Actions uses, in order to prime the same cache.

[official documentation]: https://help.github.com/en/actions/reference/virtual-environments-for-github-hosted-runners#docker-container-filesystem