Don't change directory in test/run-tests.sh
Changing directory in a shell functions changes the working directory
for the whole session and not only for the function. As such we need
to return to the previous working directory before exiting the
function. A better solution is to not change directory at all and always
specify the destination for the commands that support it.