Don't use a docker image for Tectonic
We discovered that GitHub Actions mount their own home directory when
using a docker runner ([docs]). This means that our primed cache that we
thought we could use (and the whole reason for typesetting cache.tex)
was completely ignored by Tectonic and we instead had to download
everything.
Therefore we now scrap the docker approach, instead leaning on GitHub
Actions and their caching for letting us only build Tectonic once and
then use the cached build.

This fixes #18.

[docs]: https://help.github.com/en/actions/reference/virtual-environments-for-github-hosted-runners#docker-container-filesystem