Ignore subtitle if missing
The template should respect if an author decides not to have a
subtitle. Previously it failed with inexplicable errors, originating
from inside the template (it attempts to end a line and place some
vertical space, but unless the subtitle exists the line is empty and
TeX becomes sad). This commit makes the subtitle printing
conditional.