Ignore division if missing
Not all reports specify a division and this should therefore not be a
required field by the template.

Omitting division previosly caused LaTeX fail with inexplicable
errors--it attempts to end a line and place some vertical space, but
unless the division exists the line is empty and TeX becomes sad.
This commit fixes this by not attempting to typeset division if omitted.