Add support for cover images
This commit adds support for specifying a cover image and its subtitle
using the command `\coverimage{figure-path}{figure-caption}`.

The golden files needed updating because we no longer typeset "Cover:
Description of the picture on the cover page (if applicable)", but
instead the figure-caption or nothing if omitted.

The test runner script needed updating so that it also copied over any
resources such as images etc.

This fixes #30